Magic fixes, repr, experiment with object table

+++ halolib/field.pyx
@@ -37,9 +37,9 @@ cdef extern from "string.h" nogil:
 def py_strlen(s):
     return strlen(<char*>s)
 
-# ----------------------------------------------------------------
-# Get primitive values from pointer location
 
+# Get primitive values from pointer location
+# ----------------------------------------------------------------
 def read_float32(int address): return (<float*>address)[0]
 def read_float64(int address): return (<double*>address)[0]
 def read_int8(int address):    return (<int8_t*>address)[0]
@@ -51,19 +51,18 @@ def read_uint16(int address):  return (<uint16_t*>address)[0]
 def read_uint32(int address):  return (<uint32_t*>address)[0]
 def read_uint64(int address):  return (<uint64_t*>address)[0]
 
-# ----------------------------------------------------------------
 # Assign primitive values to pointer location
-
-def write_float32(int address, float value): (<float*>address)[0] = value
-def write_float64(int address, float value): (<double*>address)[0] = value
-def write_int8(int address, int value):      (<int8_t*>address)[0] = value
-def write_int16(int address, int value):     (<int16_t*>address)[0] = value
-def write_int32(int address, int value):     (<int32_t*>address)[0] = value
-def write_int64(int address, int value):     (<int64_t*>address)[0] = value
-def write_uint8(int address, int value):     (<uint8_t*>address)[0] = value
-def write_uint16(int address, int value):    (<uint16_t*>address)[0] = value
-def write_uint32(int address, int value):    (<uint32_t*>address)[0] = value
-def write_uint64(int address, int value):    (<uint64_t*>address)[0] = value
+# ----------------------------------------------------------------
+def write_float32(int address, float value):    (<float*>address)[0] = value
+def write_float64(int address, float value):    (<double*>address)[0] = value
+def write_int8(int address,    int8_t value):   (<int8_t*>address)[0] = value
+def write_int16(int address,   int16_t value):  (<int16_t*>address)[0] = value
+def write_int32(int address,   int32_t value):  (<int32_t*>address)[0] = value
+def write_int64(int address,   int64_t value):  (<int64_t*>address)[0] = value
+def write_uint8(int address,   uint8_t value):  (<uint8_t*>address)[0] = value
+def write_uint16(int address,  uint16_t value): (<uint16_t*>address)[0] = value
+def write_uint32(int address,  uint32_t value): (<uint32_t*>address)[0] = value
+def write_uint64(int address,  uint64_t value): (<uint64_t*>address)[0] = value
 
 
 def make_field_property(node, chunk_class=None):
